Temperature microservice

+++ microservices/temperature-microservice/src/main/java/hr/fer/labpro/rassus/temperaturemicroservice/service/CurrentReadingService.java
@@ -1,7 +1,5 @@
 package hr.fer.labpro.rassus.temperaturemicroservice.service;
 
-import org.springframework.stereotype.Service;
-
 public interface CurrentReadingService {
-    Integer getCurrentTemperatureReading();
+    Double getCurrentTemperatureReading();
 }
